+++ src/contexts/useSdkConfigurationService.tsx
@@ -15,6 +15,9 @@ type SdkConfigurationServiceProviderProps = {
 const defaultConfiguration: SdkConfiguration = {
   experimental_features: {
     enable_multi_frame_capture: false,
+    motion_experiment: {
+      enabled: false,
+    },
   },
   sdk_features: {
     enable_require_applicant_consents: true,