+++ src/demo/demoUtils.ts
@@ -216,12 +216,10 @@ export const getInitSdkOptions = (): SdkOptions => {
         type: 'document',
         options: {
           useLiveDocumentCapture:
-            queryParamToValueString.useLiveDocumentCapture === 'true',
+            queryParamToValueString.useLiveDocumentCapture !== 'false',
           uploadFallback: queryParamToValueString.uploadFallback !== 'false',
           useWebcam: queryParamToValueString.useWebcam === 'true',
           documentTypes: getPreselectedDocumentTypes(),
-          showCountrySelection:
-            queryParamToValueString.oneDocWithCountrySelection === 'true',
           forceCrossDevice: queryParamToValueString.forceCrossDevice === 'true',
           requestedVariant:
             queryParamToValueString.docVideo === 'true' ? 'video' : 'standard',