+++ src/types/commons.ts
@@ -1,7 +1,13 @@
 import { EventEmitter2 } from 'eventemitter2'
 import { EnterpriseFeatures } from './enterprise'
 import { SupportedLanguages, LocaleConfig } from './locales'
-import { DocumentTypes, PoaTypes, StepConfig, StepTypes } from './steps'
+import {
+  DocumentTypes,
+  PoaTypes,
+  PrivateStepConfig,
+  StepConfig,
+  StepTypes,
+} from './steps'
 import { SdkOptions } from './sdk'
 import { UICustomizationOptions } from './ui-customisation-options'
 
@@ -28,9 +34,15 @@ const STEP_CROSS_DEVICE = 'crossDevice'
 export type ExtendedStepTypes = StepTypes | typeof STEP_CROSS_DEVICE
 export type ExtendedStepConfig =
   | StepConfig
-  | { type: typeof STEP_CROSS_DEVICE; options?: never; skip?: boolean }
+  | ({ type: typeof STEP_CROSS_DEVICE; options?: never } & PrivateStepConfig)
 
-export type CaptureMethods = 'poa' | 'document' | 'face' | 'auth' | 'data'
+export type CaptureMethods =
+  | 'poa'
+  | 'document'
+  | 'face'
+  | 'auth'
+  | 'data'
+  | 'activeVideo'
 
 export type CaptureMethodVariants = 'live' | 'html5'
 
@@ -106,6 +118,7 @@ export type ErrorNames =
   | 'FACE_VIDEO_TIMEOUT'
   | 'PROFILE_DATA_TIMEOUT'
   | 'FORBIDDEN_CLIENT_ERROR'
+  | 'GEOBLOCKED_ERROR'
   | 'GENERIC_CLIENT_ERROR'
   | 'GLARE_DETECTED'
   | 'INTERRUPTED_FLOW_ERROR'
@@ -132,7 +145,7 @@ export type documentSelectionType = {
 }
 
 export type MobileConfig = {
-  clientStepIndex?: number
+  stepIndex: number
   deviceHasCameraSupport?: boolean
   disableAnalytics?: boolean
   useWorkflow?: boolean
@@ -142,7 +155,6 @@ export type MobileConfig = {
   poaDocumentCountry?: CountryData
   language?: SupportedLanguages | LocaleConfig
   poaDocumentType?: PoaTypes
-  step?: number
   steps: StepConfig[]
   token?: string
   urls: UrlsConfig
@@ -152,6 +164,7 @@ export type MobileConfig = {
   crossDeviceClientIntroProductName?: string
   crossDeviceClientIntroProductLogoSrc?: string
   analyticsSessionUuid?: string
+  workflowRunId?: string
 }
 
 export type FormattedError = {